Rename to fix packaging issues for Melpa
Melpa requires that every package contains a library whose name that
matches the of the package.  In this case "omtose-phellack-theme.el"
already exists.

M-x load-theme offers all libraries named "*-theme.el", which are
located on located on `custom-theme-load-path' as themes, without
checking whether these libraries actually define a theme.

"omtose-phellack-theme.el" does not define a theme -- it exists to
satisfy Melpa's matching library name requirement.

"package.el" expects all libraries in a package to be located at the
top-level, so "package-build.el" (the tool used to build packages for
Melpa) moves all libraries to the top-level.

As a result the actual themes provided by this package and the metadata
file end up in the same directory.

This is not the only multi-variant theme distributed on Melpa, so we
have come up with the convention that multi-variant theme packages (and
collections of assorted themes, distributed as a single package, must be
named "-*themes".  This not only addresses the above issues, it also
makes a lot of sense: by using the plural, we communicate to the
interested user, that the package provides more than a single theme.

I was unable to contact the author and while I would normally remove
such a package from Melpa and the Emacsmirror, I have decided to fix
it and semi-maintain a fork on the Emacsorphanage, because this has
over a  hundred thousand downloads from Melpa.

Backstory:
- https://github.com/melpa/melpa/pull/4033
- https://github.com/melpa/melpa/pull/4039
- https://github.com/melpa/melpa/issues/8219#issuecomment-1464900236